优化DP求解器占用内存大小

+++ src-tauri/src/dynamic_programing_solver.rs
@@ -5,7 +5,7 @@ use std::cell::Cell;
 #[derive(Clone, Copy, Default)]
 struct SolverSlot<T> {
     value: T,
-    step: i32,
+    step: u16,
     action: Option<Actions>,
 }
